Bug Fixes, Tests, and Memory Rework
Fixed a minor bug in jal (off by one on a shift) that was having major
reprecussions (only worked 25% or so of the time)
Changed our memory array's size to mirror the venus simulator and match
the oldloop test
Changed the stack pointer so it started at the same place as venus's
Fixed the global pointer so it had a starting index, and selected the
same index venus uses
Fixed the stack pointer so that it was byte addressed, and modified the
eff_addr calculations to account for that
Removed some trivial code (e.g., & -1)
Added the ability to store and retrieve data across unaligned words.
This required rewriting sh, sw, lh, lhu, lw, and lwu.
Added string print to ecall (untested)
Added an output function as the TA's required
Converted the current manual testing harness to a simple runner class
for the demo Monday
Replaced the manual testing and visual comparison with automated JUnit5
test classes
Added test files for all instructions we had to implement
Reorganized package/file structure