More tests and bug fixes
Added tests for ecall (needs visual inspection)
Fixed a bug in ecall 4 (string printing)
Added tests for mis-aligned loads/stores
Modified one of the "random" tests to show it worked. Other 9 seemed
largely redundant so removed them
Fixed a bug that caused unimplemented instructions to be processed and
adversely affect output. Now if an unimplimented instruction is
encountered the simulator skips it and an error is printed
Code coverage of simulator is currently at 95.5%+!