Changed Memory Convention
Before we had memory get larger as you moved left across a word. However
we needed memory to get larger as you move right (since that is the
convention on how memory works). Once we corrected this, we had to
reverse the bytes in words and half-words before storage and after
retrieval to maintain little-endian memory. The most significant byte
should be in the biggest memory slot. This happened naturally before
(bigger memory slot to the left, MSB is on the left) which is why it
wasn't previously an issue.